Rename to Jellyfin conventions
