+++ Jellyfin.Plugin.PlaybackReporting.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2036
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "playback_reporting", "playback_reporting\playback_reporting.csproj", "{46521AAF-B2F0-4054-8514-F619CA3EB3E4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Jellyfin.Plugin.PlaybackReporting", "Jellyfin.Plugin.PlaybackReporting\Jellyfin.Plugin.PlaybackReporting.csproj", "{46521AAF-B2F0-4054-8514-F619CA3EB3E4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
