+++ Jellyfin.Plugin.PlaybackReporting/Data/ManagedConnection.cs
@@ -16,12 +16,9 @@ along with this program. If not, see<http://www.gnu.org/licenses/>.
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using SQLitePCL.pretty;
 
-namespace playback_reporting.Data
+namespace Jellyfin.Plugin.PlaybackReporting.Data
 {
     public class ManagedConnection : IDisposable
     {
