+++ Jellyfin.Plugin.PlaybackReporting/Data/SqliteExtensions.cs
@@ -18,11 +18,10 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using MediaBrowser.Model.IO;
 using MediaBrowser.Model.Serialization;
 using SQLitePCL.pretty;
 
-namespace playback_reporting.Data
+namespace Jellyfin.Plugin.PlaybackReporting.Data
 {
     public static class SqliteExtensions
     {
