+++ Jellyfin.Plugin.PlaybackReporting/Extensions.cs
@@ -17,7 +17,7 @@ along with this program. If not, see<http://www.gnu.org/licenses/>.
 using System.Collections.Generic;
 using MediaBrowser.Common.Configuration;
 
-namespace playback_reporting
+namespace Jellyfin.Plugin.PlaybackReporting
 {
     public static class ConfigurationExtension
     {
