+++ Jellyfin.Plugin.PlaybackReporting/Pages/breakdown_report.html
