+++ Jellyfin.Plugin.PlaybackReporting/Pages/custom_query.js
