+++ Jellyfin.Plugin.PlaybackReporting/Pages/duration_histogram_report.html
