+++ Jellyfin.Plugin.PlaybackReporting/Pages/duration_histogram_report.js
@@ -157,7 +157,7 @@ define(['libraryMenu'], function (libraryMenu) {
         // init code here
         view.addEventListener('viewshow', function (e) {
 
-            libraryMenu.setTabs('playback_reporting', 4, getTabs);
+            libraryMenu.setTabs('Jellyfin.Plugin.PlaybackReporting', 4, getTabs);
 
             require([Dashboard.getConfigurationResourceUrl('Chart.bundle.min.js')], function (d3) {
