+++ Jellyfin.Plugin.PlaybackReporting/Pages/hourly_usage_report.js
