+++ Jellyfin.Plugin.PlaybackReporting/Pages/playback_report_settings.html
