+++ Jellyfin.Plugin.PlaybackReporting/Pages/user_playback_report.js
