+++ Jellyfin.Plugin.PlaybackReporting/Pages/user_report.html
