+++ Jellyfin.Plugin.PlaybackReporting/Pages/user_report.js
