+++ Jellyfin.Plugin.PlaybackReporting/PluginConfiguration.cs
@@ -14,15 +14,12 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see<http://www.gnu.org/licenses/>.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Text;
+using MediaBrowser.Model.Plugins;
 
-namespace playback_reporting.Api
+namespace Jellyfin.Plugin.PlaybackReporting
 {
-    public class ReportDayUsage
+    public class PluginConfiguration: BasePluginConfiguration
     {
-        public string Date { get; set; }
-        public int Count { get; set; }
+
     }
 }
