+++ Jellyfin.Plugin.PlaybackReporting/ReportPlaybackOptions.cs
@@ -14,24 +14,13 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see<http://www.gnu.org/licenses/>.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace playback_reporting.Data
+namespace Jellyfin.Plugin.PlaybackReporting
 {
-    public class PlaybackInfo
+    public class ReportPlaybackOptions
     {
-        public string Id { get; set; }
-        public DateTime Date { get; set; }
-        public string UserId { get; set; }
-        public string ItemId { get; set; }
-        public string ItemType { get; set; }
-        public string ItemName { get; set; }
-        public string PlaybackMethod { get; set; }
-        public string ClientName { get; set; }
-        public string DeviceName { get; set; }
-        public int PlaybackDuration { get; set; } = 0;
+        public int MaxDataAge { set; get; } = 3;
+        public string BackupPath { set; get; }
+        public int MaxBackupFiles { set; get; } = 5;
 
     }
 }
