+++ Jellyfin.Plugin.PlaybackReporting/thumb.png
